Cleaning up test driver splitting to sub modules
